Add files via upload
We are presenting an application of the proposed capacitive-coupling-based EQS HBC-HSI link by demonstrating the real-time transmission of an audio signal from an on-body transmitter (i.e., smartphone) through the human body and a conducting structure and successful reception to an off-body receiver (i.e., wearable headset). The demonstrated audio streaming exemplifies the feasibility of medium-speed data transfer (1.44 Mbps) over an on-body-to-off-body communication link using battery-powered wearable form factor devices. The mode of communication, being digital, with a low enough bit-error rate, ensures reliable transfer of the digital data packets while maintaining the quality of the audio transmission.